Makefile: restore *.bin.pmecc rule compatibility with MPLAB X
1. Use test operators compatible with GnuWin32 (detailed below).
2. Edit pmecchead() so that the '*.bin.pmecc' rule in the Makefile will
not reuse an inadequate 'pmecc.tmp' that persisted from a different
build. That is, handle the case where pmecchead() skips writing to
'pmecc.tmp', yet returns success.

Since commit 01cd9077eff53698a4c8d4f13a044de669907b1a and release v4.0.2,
the recipe for at91bootstrap.bin.pmecc has been testing whether the
'pmecc.tmp' intermediate file exists. While the syntax of this test is
fine for a Unix-like shell, it is not supported by GnuWin32. As a
result, this test has broken the build under MPLAB X IDE for Windows,
whenever a NAND Flash is targeted.

MPLAB X IDE for Windows relies on a subset of GnuWin32.
GnuWin32's implementation of GNU make runs the recipes as native batch
files.
We can mimic this test at Command Prompt's invite:

    CMD> cd "\Program Files\Microchip\MPLABX\v6.00\gnuBins\GnuWin32\bin"
    CMD> if [ -f libiconv2.dll ]; then echo YES; fi
    -f was unexpected at this time.

The above error is the one the build fails upon.
Nevertheless, alternate forms can be used:

    CMD>[ -f libiconv2.dll ] && echo YES
    YES
    CMD>test -f libiconv2.dll && echo YES
    YES
    CMD>test -f dummy.dll && echo YES

Let's implement the latter.
Note: Windows seems to partially implement the AND-OR list specified in
IEEE Std 1003.1: the '&&' operator and a subsequent '||' operator seem
to comply, but the reverse does not.

Fixes: 01cd9077 ("scripts: addpmecchead.py: Allow binaries without PMECC header")
Reported-by: Jeba Shiny <jeba.shiny@microchip.com>
Signed-off-by: Aubin Constans <aubin.constans@microchip.com>