configs: sama5d2_xplained: Add OP-TEE SD defconfig
Add dedicated OP-TEE defconfig for sama5d2_xplained. This defconfig
was derived from sama5d2_xplainedsd_uboot_defconfig with the single
addition of "CONFIG_LOAD_OPTEE=y" and implicitly the removal of
"# CONFIG_ENTER_NWD is not set". Copyright year was updated to 2022.
"make savedfconfig" was not used to update the defconfig as it
removed few config entries. We'll deal with that later on.

There are two advantages in introducing a dedicated defconfig for OP-TEE:
1/ you can build the at91bootstrap OP-TEE binary by hand
2/ CI can test the defconfig periodically.

Signed-off-by: Tudor Ambarus <tudor.ambarus@microchip.com>