device: sam9x60: update SRAM and DDRAM's DCache settings
Fix the issue that load image from NAND flash to DDRAM is
slower when CONFIG_MMU=y and CONFIG_CACHES=y

For an image with the size 4M bytes, the load will spend %6 more time than
not using MMU/cache.
With the patch the load will save %5.3 time comparing with not using MMU/cache.

Tested-by: Xing Chen <xing.chen@microchip.com>
Signed-off-by: Tony Han <tony.han@microchip.com>