crt0_gnu.S: always pass along r4 to next stage
At least on the SAMA5D2, D3 and D4, the BootROM encodes information about
the boot source into the r4 register. AT91Bootstrap discards this
information on all SoCs except for the SAMA5D3. Fix this by storing r4
at a universally applicable location (first word of stack) and passing
it along to the next boot stage, so it can decide how to act on this.

In order to keep backwards compatibility with next stage boot firmware
that looks at the GPBR instead of r4, we keep the store to GPBR
unchanged and replace only the load with aforementioned stack top.

Acked-by: Nicolas Ferre <nicolas.ferre@microchip.com>
Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>