configs: sama5d3-eds: add board support
Previously, the sama5d3-eds was sharing the config of the
sama5d3_edxplained.  However, the size of uboot has grown
and now requires a different IMG_SIZE.  This is the only
change at this point from the sama5d3_explained config.
More cleanup my follow, but this is a pre-requisite patch
for modifying the uboot and buildroot-external repos.

Signed-off-by: Jerry Ray <jerry.ray@microchip.com>