driver: at91-qspi: bug fix qspi clock setting for sama7d65
The QSPI relies on a DLL to generate the clock required for the
Controller core and the physical interface for sama7g5, but for sama7d65
does not need it.
For sama7d65, the QSPI GCLK is a 2x clock, it must be set to 266MHz to
reach 133MHz on the data.
 .

Signed-off-by: LI BIN <bin.li@microchip.com>