Makefile: generate Plain Text Image for new SoCs
Newer SoCs have a new binary format with a header and footer added
to the raw binary, needed by the RomCode. This format is described
in the product's datasheet ("Bootstrap Image Format" chapter).

Use the SAM-BA tool to generate a Plain Text Image format,
non-secure mode, from the raw binary and add it to the build process.
The SAM-BA is checked so that user is warned if the tool is not
installed or not having the needed feature.

Signed-off-by: Nicolas Ferre <nicolas.ferre@microchip.com>