driver:pmecc: fix conflicated PMECC version check
To determine if the SOC is SAMA5D3, it checks by reading the PMERRLOC
version number, but for most of the later SOCs(9x7,sama7g5...), the
PMERRLOC version register no longer exists and is now at the end of the
HSMC register. This fix only checks CONFIG_SAMA5D3 not the version
check.

Signed-off-by: LI BIN <bin.li@microchip.com>