Makefile: add LD flags: -z noexecstack --no-warn-rwx-segments
Fix new warnings appearing with versions of GNU ld (BFD) more recent
than 2.37:
arm-linux-gnueabi-ld: warning: ./build/crt0_gnu.o: missing .note.GNU-stack section implies executable stack
arm-linux-gnueabi-ld: NOTE: This behaviour is deprecated and will be removed in a future version of the linker
arm-linux-gnueabi-ld: warning: ./build/binaries/sama7g5-sdcardboot-uboot-4.0.8.elf has a LOAD segment with RWX permissions

-z noexecstack: tells the linker to make sure that our stack is
non-executable which is what we want for such a bootloader. As we
load raw binary, this ELF annotation is not preserved anyway.

--no-warn-rwx-segments: likewise, we load raw binary and don't use
ELF segment permissions: it's safe to disable this warning.

While the first option addition is supported by most of the linkers
that we use (including gcc, MPlabX, xc32), the second option must be
"protected" by a call to the ldckflags function as it's specific to BFD.

Signed-off-by: Nicolas Ferre <nicolas.ferre@microchip.com>
Acked-by: Aubin Constans <aubin.constans@microchip.com>