driver: qspi; handling qspi flash cache coherency issue
When QSPI read with CPU with MMU/Dcache enable, the problem is in the init
 (read ID, read status, read SfDP,.) which was made with AHB access (data
keep at QSPI address 0x6000000). The first access is correct, but the others
are not because we try to read at address 0x60000000 which was already in
the cache. We must invalidate the cache line after access AHB bus.

Signed-off-by: Li Bin <bin.li@microchip.com>