Kconfig: remove BACKUP_VDDIN33 from default yes
Remove the switch of VDDBU to VDDIN33 by default. This will change the
behavior of this SFRBU switch, coming back to a "reset" situation
for sama5d2, sama7g5 and sama7d65.
The switch to VDDIN33 could only be activated if we monitor this input
power source and switch back to VDDBAT quickly in case of power loss.
This is not the general case, so use safer VDDBAT in all cases (and in
particular for our evaluation kits).

Note: Defconfig files souldn't be modified but re-generating the .config
file is needed to benefit from this change.

Signed-off-by: Nicolas Ferre <nicolas.ferre@microchip.com>