driver: sfrbu: add a function to set BU power automatically
Add the sfrbu_auto_ba_power_source() function to allow to set the
Backup Unit Power Switch to automatic/hardware mode.
This mode allows to avoid having to toggle the switch by software
and be safe for power modes requiring to keep Backup Unit powered by
VDDBAT for instance.
This mode is not the one selected by on reset, so the need for way
to configure it.

Signed-off-by: Nicolas Ferre <nicolas.ferre@microchip.com>