driver:mci: add set block count command
CMD12 has been used to stop multiple-block read. however, CMD12
is timing dependent and it is difficult to control timing to
issue CMD12 at exact timing. As UHS104 card has large delay
between clock and data, CMD23 is useful for the host to stop
multiple read operation instead of CMD12. Host is not necessary
to control timing of CMD12.

Co-authored-by: Xing Chen <xing.chen@microchip.com>

Signed-off-by: Li Bin <bin.li@microchip.com>