Makefile: Add boot version support for sama7d65
Add the possibility from Kconfig to set the version of the binary while generating
the bootable image. This feature is Referred as "Bootstrap Version" in the datasheet
and used for Dual Boot or Anti Roll-Back.
Not to be mixed up with at91bootstrap software version number which is
the version of the project: they can be distinct.

Co-developed-by: Jeba Shiny <jeba.shiny@microchip.com>
Signed-off-by: Jeba Shiny <jeba.shiny@microchip.com>
Signed-off-by: Li Bin <bin.li@microchip.com>