include:securam: shift SECURAM for backup mode
In the ed48d7a96f1 commit, SECURAM was shifted by 0x800 to account
for ROM code in emulation mode, which is incorrect. When using
emulated OTP, the OTP packet is copied to the beginning of the
SECURAM memory section (starting at offset 0x800). This overwrites
BCP.EMUL and after a reset it will not be possible to boot again
(except for SD booting). In backup mode, if SECURAM is offset
from 0xe0000800, kernel->PM will overwrite BCP.EMUL when kernel is
 loaded. so not only is it not necessary to shift SECURAM by 0x800,
 but also not necessary to shift SECURAM by 0x800 in Linux PM.

Signed-off-by: Li Bin <bin.li@microchip.com>