driver:dram: fix some dram configuration issues
Fix ZPROG[7:0] Impedance Divide Ratio, the value is 13 for 34Ohms.
Fix DDR3 driver impedance value in KCONFIG, it support 34 and 40 ohms
 only.
To be on the safe side, the AS4C256M16D3LC-10BIN should be configured
as an 1866M, even though the 1600K works, but the risk is that there
is a difference in tRP and tRCD timing between the 1600K and the 1866M.
Set the speed bin to 1866 instead of 1333 for MT52L256M32D1PF_107 for
performace.

Signed-off-by: Li Bin <bin.li@microchip.com>
Signed-off-by: Hari Prasath Gujulan Elango <hari.prasathge@microchip.com>