Kconfig: introduce fast boot configurations
Create a new 'fast-boot' folder with all fast boot related header
files, Kconfigs and libraries in future.
Add a CONFIG_FAST_BOOT in Config.in.nvm. When the user selects which
non-volatile memory to load from, the selected non-volatile memory
will be used to store/load the snapshot image for fast booting if
 the FAST_BOOT option is also enabled.
Create a separate Config.in.fastboot to hold fastboot related options,
including whether or not to encrypt the Snapshot, the choice of
kernel revision, memory management related options, the entry
point for the Snapshot image, or the partition of the SD card
used for fast booting.

Signed-off-by: Li Bin <bin.li@microchip.com>