driver:at91-qspi:Fix qspi memcpy for octa mode
The existing QSPI driver can successfully configure the device to
8-8-8 mode or 8D-8D-8D mode and load images in that mode via memory
access. However, when Octa DDR mode is enabled, it failed if
access register.
The function qspi_memcpy() always accesses the data using the 8-bit
data width, which will not read the correct data when DTR is enabled.
Adding the DTR parameter in qspi_memcpy(), when DTR is true, the data
will be accessed using the full 16-bit data width, otherwise the data
will be accessed using the full 8-bit data width.

Signed-off-by: Li Bin <bin.li@microchip.com>