driver: fix DDR2 W9751G6NB timing for sam9x7 sip
The number of tRC cycles is between 0 and 15. When tRC is greater than
or equal to 60 ns, the maximum supported frequency is 250 MHz. When
the frequency is higher than 200 MHz, an offset of one is added
automatically. In such case, tRC must be set to 15 to match the value.
If the frequency is higher than 200MHz, the CAS latency must be set
higher than 3. For sam9x7, the frequency is 266MHz, the CAS should be 5.
Adjust the DDR2 timings(tRCD, tRP, tRC and tRAS) with DDR2-800D @CL=5.

Signed-off-by: Xing Chen <xing.chen@microchip.com>