Makefile: remove extra REVISION / REV variables
This REVISION and associated REV makefile variales were used for "rcx"
type of values. We can directly use VERSION string as we do for
numerous releases.
The addition of "export" to that VERSION variable and its use in
menuconfig makes the old REVISION / REV variables harmful as there could
be a difference in display of the full string.
Remove it.

Signed-off-by: Nicolas Ferre <nicolas.ferre@microchip.com>